Only Play Music for Player Scenes
Change the ostim_start behaviour to only play songs when the player is involved, and change the ostim_end behaviour to only end songs if the player is not active.

This should cover most cases of npcs' scenes overlapping with the player's (eg. music won't start for npc scenes, and npc scenes ending won't end the music during a player scene).